TST: Consolidate all array construction tests
