+++ numpy/typing/tests/data/reveal/array_constructors.py
@@ -36,3 +36,7 @@ reveal_type(np.require(B, requirements="W"))  # E: SubClass
 reveal_type(np.require(B, requirements="A"))  # E: SubClass
 reveal_type(np.require(C))  # E: ndarray
 
+reveal_type(np.linspace(0, 10))  # E: numpy.ndarray
+reveal_type(np.linspace(0, 10, retstep=True))  # E: Tuple[numpy.ndarray, numpy.inexact]
+reveal_type(np.logspace(0, 10))  # E: numpy.ndarray
+reveal_type(np.geomspace(1, 10))  # E: numpy.ndarray
